Renders all profile sections as livewire cards
- rename component to ProfileDataCard
- for consistency and code cleanliness, renders all sections w/the livewire component, not just paginated ones
- consolidate section empty check in component
- only pass information section from controller to show view; the rest are rendered in the component to avoid excess DB queries.
- improve HTML semantic roles: sections as `<section>`, section links as `<nav>`
