+++ resources/views/profiles/show.blade.php
@@ -84,84 +84,35 @@
 			</div>
 		</div>
 	</div>
-	<div id="links" class="container links">
+	<nav id="links" class="container links" aria-label="profile sections">
 		<ul>
-			@if(!$preparations->isEmpty())<li><a href="#preparation">Professional Preparation</a></li>@endif
-			@if(!$research_areas->isEmpty())<li><a href="#areas">Research Areas</a></li>@endif
-			@if(!$publications->isEmpty())<li><a href="#publications">Publications</a></li>@endif
-			@if(!$appointments->isEmpty())<li><a href="#appointments">Appointments</a></li>@endif
-			@if(!$awards->isEmpty())<li><a href="#awards">Awards</a></li>@endif
-			@if(!$projects->isEmpty())<li><a href="#projects">Projects</a></li>@endif
-			@if(!$presentations->isEmpty())<li><a href="#presentations">Presentations</a></li>@endif
-			@if(!$additionals->isEmpty())<li><a href="#additional">Additional Information</a></li>@endif
-			@if(!$news->isEmpty())<li><a href="#news">News</a></li>@endif
-			@if(!$activities->isEmpty())<li><a href="#activities">Activities</a></li>@endif
-			@if(!$affiliations->isEmpty())<li><a href="#affiliations">Affiliations</a></li>@endif
-			@if(!$support->isEmpty())<li><a href="#funding">Support</a></li>@endif
+			@if($profile->preparation()->exists())<li><a href="#preparation">Professional Preparation</a></li>@endif
+			@if($profile->areas()->exists())<li><a href="#areas">Research Areas</a></li>@endif
+			@if($profile->publications()->exists())<li><a href="#publications">Publications</a></li>@endif
+			@if($profile->appointments()->exists())<li><a href="#appointments">Appointments</a></li>@endif
+			@if($profile->awards()->exists())<li><a href="#awards">Awards</a></li>@endif
+			@if($profile->projects()->exists())<li><a href="#projects">Projects</a></li>@endif
+			@if($profile->presentations()->exists())<li><a href="#presentations">Presentations</a></li>@endif
+			@if($profile->additionals()->exists())<li><a href="#additionals">Additional Information</a></li>@endif
+			@if($profile->news()->public()->exists())<li><a href="#news">News</a></li>@endif
+			@if($profile->activities()->exists())<li><a href="#activities">Activities</a></li>@endif
+			@if($profile->affiliations()->exists())<li><a href="#affiliations">Affiliations</a></li>@endif
+			@if($profile->support()->exists())<li><a href="#funding">Support</a></li>@endif
 		</ul>
-	</div>
+	</nav>
 	<div class="container card-columns main_areas">
-	@if(!$preparations->isEmpty() || $editable)
-		<div class="card">
-			<h3 id="preparation"><i class="fas fa-graduation-cap" aria-hidden="true"></i> Professional Preparation @if($editable)<a class="btn btn-primary btn-sm" href="{{ route('profiles.edit', [$profile->slug, 'preparation']) }}"><i class="fas fa-edit"></i> Edit</a>@endif</h3>
-			@foreach($preparations as $prep)
-				<div class="entry">
-					{{$prep->degree}} @if($prep->major)- {{$prep->major}}@endif
-					<br>
-					<strong>{{$prep->institution}}</strong>@if($prep->year) - {{$prep->year}}@endif
-				</div>
-			@endforeach
-		</div>
-	@endif
-	@if(!$research_areas->isEmpty() || $editable)
-			<div class="card">
-				<h3 id="areas"><i class="fas fa-flask" aria-hidden="true"></i> Research Areas @if($editable)<a class="btn btn-primary btn-sm" href="{{ route('profiles.edit', [$profile->slug, 'areas']) }}"><i class="fas fa-edit"></i> Edit</a>@endif</h3>
-				@foreach($research_areas as $area)
-					@if($area->url)
-						<h5><a href="{{$area->url}}">{{$area->title}} <i class="fas fa-link" aria-hidden="true"></i></a></h5>
-					@else
-						<h5>{{$area->title}}</h5>
-					@endif
-					{!! Purify::clean($area->description) !!}
-				@endforeach
-			</div>
-	@endif
-	@if($paginated)
-		<livewire:paginated-data :editable="$editable" :profile="$profile" data_type="publications">
-		<livewire:paginated-data :editable="$editable" :profile="$profile" data_type="awards">
-		<livewire:paginated-data :editable="$editable" :profile="$profile" data_type="appointments">
-		<livewire:paginated-data :editable="$editable" :profile="$profile" data_type="projects">
-		<livewire:paginated-data :editable="$editable" :profile="$profile" data_type="presentations">
-		<livewire:paginated-data :editable="$editable" :profile="$profile" data_type="additionals">
-		<livewire:paginated-data :editable="$editable" :profile="$profile" data_type="news">
-	@else
-		@include('livewire/profile-data/publications', ['data' => $publications, 'paginated' => $paginated])
-		@include('livewire/profile-data/awards', ['data' => $awards, 'paginated' => $paginated])
-		@include('livewire/profile-data/appointments', ['data' => $appointments, 'paginated' => $paginated])
-		@include('livewire/profile-data/projects', ['data' => $projects, 'paginated' => $paginated])
-		@include('livewire/profile-data/presentations', ['data' => $presentations, 'paginated' => $paginated])
-		@include('livewire/profile-data/additionals', ['data' => $additionals, 'paginated' => $paginated])
-		@include('livewire/profile-data/news', ['data' => $news, 'paginated' => $paginated])
-	@endif
-	@if(!$activities->isEmpty() || $editable)
-		<div class="card">
-			<h3 id="activities"><i class="fas fa-chart-line" aria-hidden="true"></i> Activities @if($editable)<a class="btn btn-primary btn-sm" href="{{ route('profiles.edit', [$profile->slug, 'activities']) }}"><i class="fas fa-edit"></i> Edit</a>@endif</h3>
-			@foreach($activities as $activity)
-				<div class="entry">
-					<h5>{{$activity->title}}</h5>
-					{!! Purify::clean($activity->description) !!}
-					@if($activity->start_date)[{{$activity->start_date}}&ndash;{{$activity->end_date}}] @endif
-				</div>
-			@endforeach
-		</div>
-	@endif
-	@if($paginated)
-		<livewire:paginated-data :editable="$editable" :profile="$profile" data_type="affiliations">
-		<livewire:paginated-data :editable="$editable" :profile="$profile" data_type="support">
-	@else
-		@include('livewire/profile-data/affiliations', ['data' => $affiliations, 'paginated' => $paginated])
-		@include('livewire/profile-data/support', ['data' => $support, 'paginated' => $paginated])
-	@endif
+		<livewire:profile-data-card :editable="$editable" :profile="$profile" :paginated="false" data_type="preparation">
+		<livewire:profile-data-card :editable="$editable" :profile="$profile" :paginated="false" data_type="areas">
+		<livewire:profile-data-card :editable="$editable" :profile="$profile" :paginated="$paginated" data_type="publications">
+		<livewire:profile-data-card :editable="$editable" :profile="$profile" :paginated="$paginated" data_type="awards">
+		<livewire:profile-data-card :editable="$editable" :profile="$profile" :paginated="$paginated" data_type="appointments">
+		<livewire:profile-data-card :editable="$editable" :profile="$profile" :paginated="$paginated" data_type="projects">
+		<livewire:profile-data-card :editable="$editable" :profile="$profile" :paginated="$paginated" data_type="presentations">
+		<livewire:profile-data-card :editable="$editable" :profile="$profile" :paginated="$paginated" data_type="additionals">
+		<livewire:profile-data-card :editable="$editable" :profile="$profile" :paginated="$paginated" :public_filtered="true" data_type="news">
+		<livewire:profile-data-card :editable="$editable" :profile="$profile" :paginated="false" data_type="activities">
+		<livewire:profile-data-card :editable="$editable" :profile="$profile" :paginated="$paginated" data_type="affiliations">
+		<livewire:profile-data-card :editable="$editable" :profile="$profile" :paginated="$paginated" data_type="support">
 	</div>
 </div>
 @stop
